Support syntax highlighting
In general, try to match the existing behaviour. There are some small
differences in behaviour here, mostly due to parsing changes, but not
much of note. I also added a couple of colours to the dark theme that
were missing, including for the doctype tag and special (e.g. regex)
strings.