Upgrade CodeMirror from v5.39.0 to v6, and delete Screen reader mode (#591)
Main changes:

* Upgrade from CodeMirror v5.39.0 to v6.
* Update all of the file type handlers. Syntax highlighting will look different.
* Delete Screen reader mode option because CM6 has better screen reader support.
* Add a build step to install the CodeMirror dependencies from npm.

Other changes:

* Make Find's Previous result and Next result buttons only appear when there are search results.
* Fix aria labels for buttons in title bar.
* Remove "default" theme, it's only used up to M103.

======

* Add a way to build the text-app with codemirror 6

And begin implementing the migration from 5 to 6.
Also disable many things that stopped working.

* Start migration to CMv6

- Remove textarea flow (screen-reader option is left in place for now),
  this will be going away
-- Remove unified session logic, just store the EditorState directly
- Fix passing extensions to CM
- Other misc..

* wire up search

* Fix search imports

Also fix save file

* Fix up search behaviour more

- Fix current match index calculation
- Clear selection before doing searches so we can more easily do
  incremental search (i.e. don't need to search back first).
- Clear selection when leaving the search box
- Tidy code a bit. Split clear() into activate()/deactivate(), which
  lets us open the (hidden) search panel in activate() instead of find.

* Add wrapLines setting

And make it the settings apply to the new tab after switching tabs.
Without `applyAllSettings`, the setting change only applies to the
current tab.

Modify the setTabSize code a bit.

Also add "indentWithTab", which makes tab indent instead of moving focus
outside the editor.

* Make show line numbers setting work again

The logic is the same as wrap lines.

* Notify tab on document change for modified file tracking

* Support dark theme and match previous colors better

* Fix suggested file name when saveAs

It broke some commits ago due to the session_ refactor

* Tidy up theming code

- Remove "default" theme, it's been replaced by the device theme since
  M103.
- Use CSS variables for most color configurations instead of having
  separate rules for light and dark colors. Use the same rules for the
  editor view's theme.

* Make Tab and Shift-Tab behave like CM5 Text app

CodeMirror.commands.indentSelection made Tab and Shift-Tab perform the
indentMore and indentLess commands. But the previous version of the text
app was different:

 * When Tab is pressed:
     * If something is selected: It does indentMore.
     * When tabs to spaces is disabled, it inserts a tab character.
     * When tabs to spaces is enabled, it inserts the number of spaces
        a tab character would occupy at that position.
 * When Shift-Tab is pressed: It auto-indents the selection.

 Auto-indent doesn't seem to be working here yet though.

* Show underlines for matched brackets

* Fix setting font size

Also explicitly set line-height to 1.2 (defaults to 1.4) to match
previous behaviour.

* Remove old (CMv5) styles for matching brackets

* Clean up EditorCodeMirror

- Remove unused enable/disable
- Implement focus(), this is used when switching tabs

* Update comment about line separators

* Add language modes depending on the file extension

This makes commands such as toggleComment and indentSelection (auto
indent) work for supported file extensions.

Syntax highlighting doesn't work yet.

* Remove screen reader mode

- Remove the setting to enable screen reader mode
- Remove logic for switching editors (the textarea editor is already
  removed)

* Remove no longer needed strings

Screen reader mode and the default theme are no longer available.
- screenReaderModeSetting, screenReaderModeHelp
- defaultThemeOption

* Remove functionality to disable settings (from screen reader mode)

* Resolve "default" theme in settings.js instead of app.js

Currently the real theme value is only known by app.js (where the
color-scheme media query is), so EditorCodeMirror will not be able to
correctly resolve the default theme. Move this logic to settings.js
(into Settings.get) to fix.

* Fix compartment variable names, some where named as "component"

* Prevent tabbing to editor when dialogs are open

* Set aria labels for buttons

These don't get read out properly currently, e.g. the minimize button is
"remove" and the maximize button is "crop_square". Set `aria-label` in
addition to `title` (tooltip) to fix.

* Remove some references to CMv5 files and fix build.py

* Support syntax highlighting

In general, try to match the existing behaviour. There are some small
differences in behaviour here, mostly due to parsing changes, but not
much of note. I also added a couple of colours to the dark theme that
were missing, including for the doctype tag and special (e.g. regex)
strings.

* Set charset=utf-8 on codemirror.next.bin.js to fix selection

Terser replaces a "\uffff" in codemirror with a string containing a
literal utf-8 U+FFFF. This breaks highlighting multiple lines then
deleting it by typing. Fix by loading the script as utf-8 (could
alternatively set `terser({format: {ascii_only: true}})`).

* Make the smart indent setting work

Match the CM5 behavior:
Whether to use the context-sensitive indentation that the mode provides
(or just indent the same as the line before). Defaults to true.

* Fix print styles - replace CMv5 with CMv6 styles

* Fix modeAutoSet when the file name changes

A way to trigger this is to "save as" your file with a different file
extension.

* Make search navigation buttons only appear if there are search results

This fixes two issues:

 * It was possible to tab to the buttons even when they were not visible
 * They were visible even when there were no search results

Also remove the width transition. It didn't appear to be working.
And I think instant transitions are better here because the change
between 0 search results and more than 0 can happen quickly and often.

* Add support for more languages

Most of these should be identical to before as they use the legacy mode.
- Removed xquery, surely no one uses this..
- Couldn't find ocaml mode, so removed.
- Rust using a non-legacy extension
- SQL using a non-legacy extension. Seems like it was broken previously.
- Added missing colours for diff types.

* Update readme to have more detail about prebuild codemirror

---------

Co-authored-by: Timothy Loh <timloh@chromium.org>