Fix cursor type when hovering text to be the I-beam (#596)
Chrome Apps' default CSS styling sets `cursor:default` on body. CMv6
doesn't override this so we end up with the normal pointer when hovering
text. Manually set `cursor:auto` on the text area of the editor so we
get the correct cursor on text, and match previous versions of the Text
App so blank space below the last line of a document remain with the
default pointer.