Optimize find performance (#602)
The current find impl is much slower (~30x on large documents) than the
previous one, mainly due to using the SearchCursor to count matches
instead of manually counting. It's preferable to use SearchCursor as it
gives us confidence our counting will match CM's counting, but for
performance revert back to manually counting.

There is still an off-by-one issue present with overlapping matches
(e.g. "a|aaa", search "aa" and go to previous match), but this was
already present in the previous implementation.